# backend/scratch/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AIHire Recruitment Platform",
    "description": "Backend API service for the AI-powered Recruitment Platform.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/offers/generate": {
      "post": {
        "tags": [
          "Offer Letter Management"
        ],
        "summary": "Generate Offer Draft",
        "description": "Generates a draft offer letter, assigns revision version, and renders PDF.",
        "operationId": "generate_offer_draft_offers_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfferCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/offers/{offer_id}/send": {
      "post": {
        "tags": [
          "Offer Letter Management"
        ],
        "summary": "Send Offer",
        "description": "Sends the generated offer letter PDF via email and sets status to Sent.",
        "operationId": "send_offer_offers__offer_id__send_post",
        "parameters": [
          {
            "name": "offer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Offer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/offers/candidate/{candidate_id}": {
      "get": {
        "tags": [
          "Offer Letter Management"
        ],
        "summary": "Get Candidate Offers",
        "description": "Lists all offer letters received by a specific candidate.",
        "operationId": "get_candidate_offers_offers_candidate__candidate_id__get",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Candidate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OfferResponse"
                  },
                  "title": "Response Get Candidate Offers Offers Candidate  Candidate Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/offers/{offer_id}/accept": {
      "post": {
        "tags": [
          "Offer Letter Management"
        ],
        "summary": "Accept Offer",
        "description": "Marks the offer letter as Accepted by the candidate.",
        "operationId": "accept_offer_offers__offer_id__accept_post",
        "parameters": [
          {
            "name": "offer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Offer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/offers/{offer_id}/reject": {
      "post": {
        "tags": [
          "Offer Letter Management"
        ],
        "summary": "Reject Offer",
        "description": "Marks the offer letter as Rejected by the candidate.",
        "operationId": "reject_offer_offers__offer_id__reject_post",
        "parameters": [
          {
            "name": "offer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Offer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/offers/{offer_id}/joined": {
      "post": {
        "tags": [
          "Offer Letter Management"
        ],
        "summary": "Mark Candidate Joined",
        "description": "Recruiter marks the candidate as Joined.",
        "operationId": "mark_candidate_joined_offers__offer_id__joined_post",
        "parameters": [
          {
            "name": "offer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Offer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/offers/{offer_id}/hired": {
      "post": {
        "tags": [
          "Offer Letter Management"
        ],
        "summary": "Mark Candidate Hired",
        "description": "Recruiter marks the candidate as Hired (final step).",
        "operationId": "mark_candidate_hired_offers__offer_id__hired_post",
        "parameters": [
          {
            "name": "offer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Offer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/offers/{offer_id}/no-show": {
      "post": {
        "tags": [
          "Offer Letter Management"
        ],
        "summary": "Mark Candidate No Show",
        "description": "Recruiter marks the candidate as No Show.",
        "operationId": "mark_candidate_no_show_offers__offer_id__no_show_post",
        "parameters": [
          {
            "name": "offer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Offer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/offers/{offer_id}/revoke": {
      "post": {
        "tags": [
          "Offer Letter Management"
        ],
        "summary": "Revoke Offer",
        "description": "Recruiter revokes the active offer (sets status to Expired).",
        "operationId": "revoke_offer_offers__offer_id__revoke_post",
        "parameters": [
          {
            "name": "offer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Offer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/offers/{offer_id}/download": {
      "get": {
        "tags": [
          "Offer Letter Management"
        ],
        "summary": "Download Offer Pdf",
        "description": "Downloads the offer letter PDF file.",
        "operationId": "download_offer_pdf_offers__offer_id__download_get",
        "parameters": [
          {
            "name": "offer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Offer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OfferCreate": {
        "properties": {
          "candidate_id": {
            "type": "integer",
            "title": "Candidate Id"
          },
          "job_id": {
            "type": "integer",
            "title": "Job Id"
          },
          "recruiter_id": {
            "type": "integer",
            "title": "Recruiter Id"
          },
          "hr_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hr Id"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "position_title": {
            "type": "string",
            "title": "Position Title"
          },
          "department": {
            "type": "string",
            "title": "Department"
          },
          "employment_type": {
            "type": "string",
            "title": "Employment Type"
          },
          "package_amount": {
            "type": "string",
            "title": "Package Amount"
          },
          "joining_date": {
            "type": "string",
            "format": "date",
            "title": "Joining Date"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "reporting_manager": {
            "type": "string",
            "title": "Reporting Manager"
          },
          "offer_expiry_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offer Expiry Date"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "candidate_id",
          "job_id",
          "recruiter_id",
          "company_name",
          "position_title",
          "department",
          "employment_type",
          "package_amount",
          "joining_date",
          "location",
          "reporting_manager"
        ],
        "title": "OfferCreate"
      },
      "OfferResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "candidate_id": {
            "type": "integer",
            "title": "Candidate Id"
          },
          "candidate_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Code"
          },
          "job_id": {
            "type": "integer",
            "title": "Job Id"
          },
          "recruiter_id": {
            "type": "integer",
            "title": "Recruiter Id"
          },
          "hr_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hr Id"
          },
          "offer_reference": {
            "type": "string",
            "title": "Offer Reference"
          },
          "offer_version": {
            "type": "integer",
            "title": "Offer Version"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "candidate_name": {
            "type": "string",
            "title": "Candidate Name"
          },
          "position_title": {
            "type": "string",
            "title": "Position Title"
          },
          "department": {
            "type": "string",
            "title": "Department"
          },
          "employment_type": {
            "type": "string",
            "title": "Employment Type"
          },
          "package_amount": {
            "type": "string",
            "title": "Package Amount"
          },
          "joining_date": {
            "type": "string",
            "format": "date",
            "title": "Joining Date"
          },
          "joined_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Joined Date"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "reporting_manager": {
            "type": "string",
            "title": "Reporting Manager"
          },
          "offer_status": {
            "type": "string",
            "title": "Offer Status"
          },
          "candidate_response": {
            "type": "string",
            "title": "Candidate Response"
          },
          "joining_status": {
            "type": "string",
            "title": "Joining Status"
          },
          "response_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Date"
          },
          "offer_pdf_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offer Pdf Path"
          },
          "offer_expiry_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offer Expiry Date"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "candidate_id",
          "job_id",
          "recruiter_id",
          "offer_reference",
          "offer_version",
          "company_name",
          "candidate_name",
          "position_title",
          "department",
          "employment_type",
          "package_amount",
          "joining_date",
          "location",
          "reporting_manager",
          "offer_status",
          "candidate_response",
          "joining_status",
          "created_at",
          "updated_at"
        ],
        "title": "OfferResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}